# garden-app/api/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    This is an OpenAPI Specification for https://github.com/calvinmclean/automated-garden/tree/main/garden-app

    It allows managing and interacting with Gardens and Plants for this application.

  title: Garden App - OpenAPI
  version: 1.0.0
...
paths:
  /gardens:
    post:
      tags:
        - gardens
      summary: Create a new Garden
      description: Create a new Garden resource.
      operationId: createGarden
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenResponse"
        "400":
          description: Bad Request
      requestBody:
        description: Create a new Garden
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGardenRequest"
    get:
      tags:
        - gardens
      summary: Get all Gardens
      description: Query for a list of all Gardens. Optionally include end-dated Gardens.
      operationId: getAllGardens
      parameters:
        - name: end_dated
          in: query
          description: whether or not to include end-dated Gardens in the response
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllGardensResponse"
        "400":
          description: Bad Request
  /gardens/{gardenID}:
    get:
      tags:
        - gardens
      summary: Get a Garden
      description: Get details of a Garden.
      operationId: getGarden
      parameters:
        - $ref: "#/components/parameters/GardenID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenResponse"
        "400":
          description: Bad Request
    patch:
      tags:
        - gardens
      summary: Update/Edit a Garden
      description: Update/Edit a Garden. Only certain fields of a Garden are editable.
      operationId: updateGarden
      parameters:
        - $ref: "#/components/parameters/GardenID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenResponse"
        "400":
          description: Bad Request
      requestBody:
        description: Update/Edit a new Garden
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGardenRequest"
    delete:
      tags:
        - gardens
      summary: End-date a Garden
      description: End-date a Garden. This allows deleting without actually losing the resource data. Once a Garden is end-dated, certain resources will not be available such as health and action.
      operationId: endDateGarden
      parameters:
        - $ref: "#/components/parameters/GardenID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GardenResponse"
        "400":
          description: Bad Request
  /gardens/{gardenID}/action:
    post:
      tags:
        - gardens
      summary: Execute action on a Garden
      description: This endpoint allows executing certain actions on a Garden. This will actually interact with the `garden-controller` rather than just manipulating resources.
      operationId: gardenAction
      parameters:
        - $ref: "#/components/parameters/GardenID"
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
      requestBody:
        description: Execute action on a Garden
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GardenAction"
...
  /gardens/{gardenID}/plants/{plantID}:
    get:
      tags:
        - plants
      summary: Get a Plant
      description: Get details of a Plant.
      operationId: getPlant
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/PlantID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantResponse"
        "400":
          description: Bad Request
    patch:
      tags:
        - plants
      summary: Update/Edit a Plant
      description: Update/Edit a Plant. Only certain fields of a Plant are editable.
      operationId: updatePlant
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/PlantID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantResponse"
        "400":
          description: Bad Request
      requestBody:
        description: Update/Edit a new Plant
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlantRequest"
    delete:
      tags:
        - plants
      summary: End-date a Plant
      description: End-date a Plant. This allows deleting without actually losing the resource data. Once a Plant is end-dated, certain resources will not be available such as health and action.
      operationId: endDatePlant
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/PlantID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantResponse"
        "400":
          description: Bad Request
...
components:
  parameters:
    GardenID:
      name: gardenID
      in: path
      description: ID of Garden resource for this request
      required: true
      schema:
        $ref: "#/components/schemas/xid"
    PlantID:
      name: plantID
      in: path
      description: ID of Plant resource for this request
      required: true
      schema:
        $ref: "#/components/schemas/xid"
...
  schemas:
    xid:
      type: string
      description: 20 character unique identifier used throughout this API
      externalDocs:
        url: https://github.com/rs/xid
      pattern: "[0-9a-v]{20}"
      example: 9m4e2mr0ui3e8a215n4g
    link:
      type: object
      description: used for JSON Hypermedia to link to additional resources
      properties:
        ref:
          type: string
          description: one-word explanation of what this link is for
        href:
          type: string
          description: relative path for the specified resource
    UpdatePlantRequest:
      type: object
      description: This allows updating/editing a Plant resource
      allOf:
        - $ref: "#/components/schemas/Plant"
        - properties:
            created_at:
              type: string
              format: date-time
              description: the date-time when the Plant was originally created
...
    PlantResponse:
      type: object
      description: This is the response object for Plants that contains extra information only available on Plants that are created
      allOf:
        - $ref: "#/components/schemas/Plant"
        - properties:
            id:
              $ref: "#/components/schemas/xid"
            created_at:
              type: string
              format: date-time
              description: the date-time when the Plant was originally created
            end_date:
              type: string
              format: date-time
              description: the date-time when the Plant was deleted/removed
            next_water_time:
              type: string
              description: time of the next scheduled watering
              format: date-time
            moisture:
              type: number
              format: float
              description: moisture percentage of a Plant with a soil moisture sensor
            links:
              type: array
              items:
                $ref: "#/components/schemas/link"
              example:
                - rel: self
                  href: /gardens/c22tmvucie6n6gdrpal0/plants/c3ucvu06n88pt1dom670
                - rel: garden
                  href: /gardens/c22tmvucie6n6gdrpal0
                - rel: zone
                  href: /gardens/c22tmvucie6n6gdrpal0/zones/c3ucvu06n88pt1dom670
                - rel: action
                  href: /gardens/c22tmvucie6n6gdrpal0/plants/c3ucvu06n88pt1dom670/action
      required:
        - id
        - created_at
        - links
        - name
    Plant:
      type: object
      description: |
        This represents a single Plant, or more generally a "watered resource," attached to the `garden-controller`.
        It has a `plant_position` field to inform the `garden-controller` of which pump or valve to control for watering.
      properties:
        name:
          type: string
          description: this is the name of the Plant
          example: lettuce
        zone_id:
          type: object
          description: ID referencing the Zone that this Plant is a part of
          allOf:
            - $ref: "#/components/schemas/xid"
        details:
          type: object
          description: this holds various additional information about a Plant that might be useful to remember, but is not related to core functionality of this application.
          properties:
            description:
              type: string
              description: a general description of a Plant
              example: nutritious leafy green
            notes:
              type: string
              description: general notes about the Plant
              example: grown from seed and planted about 6 inches apart
            time_to_harvest:
              type: string
              description: estimate of time it takes Plant to grow
              example: 70 days
            count:
              type: integer
              description: number of actual plants that would be managed
              example: 6
...
    GardenAction:
      type: object
      description: collects all the possible actions for a Garden into a single struct so these can easily be received as one request
      properties:
        light:
          $ref: "#/components/schemas/LightAction"
        stop:
          $ref: "#/components/schemas/StopAction"
    LightAction:
      type: object
      description: turn a Garden's light on or off. If left blank or empty, it will toggle the current state
      properties:
        state:
          $ref: "#/components/schemas/LightState"
        for_duration:
          type: string
          format: duration
          description: duration string to determine how long to delay turning the light on. Only allowed to be used with state=OFF
          example: 14h
    StopAction:
      type: object
      description: instruct a Garden to stop watering
      properties:
        all:
          type: boolean
          description: whether or not the Garden's watering queue should be cleared in addition to stopping current watering
    LightState:
      type: string
      enum: [ON, OFF]
      description: state that will be set by the next action
    CreateGardenRequest:
      type: object
      description: This allows creating a Garden resource
      allOf:
        - $ref: "#/components/schemas/Garden"
      required:
        - name
    UpdateGardenRequest:
      type: object
      description: This allows updating/editing a Garden resource
      allOf:
        - $ref: "#/components/schemas/Garden"
        - properties:
            created_at:
              type: string
              format: date-time
              description: the date-time when the Garden was originally created
    GardenResponse:
      type: object
      description: This is the response object for Gardens that contains extra information only available on Gardens that are created
      allOf:
        - $ref: "#/components/schemas/Garden"
        - properties:
            id:
              $ref: "#/components/schemas/xid"
            created_at:
              type: string
              format: date-time
              description: the date-time when the Garden was originally created
            end_date:
              type: string
              format: date-time
              description: the date-time when the Garden was deleted/removed
            next_light_action:
              type: object
              description: time and state for the next scheduled LightAction
              properties:
                time:
                  type: string
                  format: date-time
                  description: date-time of the next action
                state:
                  $ref: "#/components/schemas/LightState"
            num_plants:
              type: integer
              description: the number of active (not end-dated) Plants in this Garden
              example: 1
              minimum: 0
            plants:
              description: link specifically for the collection of Plants
              allOf:
                - $ref: "#/components/schemas/link"
              example:
                rel: collection
                href: /gardens/c6corm5mk1u53v1dchq0/plants
            zones:
              description: link specifically for the collection of Zones
              allOf:
                - $ref: "#/components/schemas/link"
              example:
                rel: collection
                href: /gardens/c6corm5mk1u53v1dchq0/zones
            links:
              type: array
              items:
                $ref: "#/components/schemas/link"
              example:
                - rel: self
                  href: /gardens/c6corm5mk1u53v1dchq0
                - rel: health
                  href: /gardens/c6corm5mk1u53v1dchq0/health
                - rel: plants
                  href: /gardens/c6corm5mk1u53v1dchq0/plants
                - rel: zones
                  href: /gardens/c6corm5mk1u53v1dchq0/zones
                - rel: action
                  href: /gardens/c6corm5mk1u53v1dchq0/action
      required:
        - id
        - created_at
        - plants
        - zones
        - links
    AllGardensResponse:
      type: object
      description: List of all Gardens
      properties:
        gardens:
          type: array
          items:
            $ref: "#/components/schemas/GardenResponse"
    Garden:
      type: object
      description: A Garden corresponds to a `garden-controller` in the real-world and contains all the necessary information for interacting with it
      properties:
        name:
          type: string
          description: A descriptive name for the Garden
          example: Garden
        topic_prefix:
          type: string
          description: |
            This is the prefix of the MQTT topic for sending messages to the controller, so it is recommended to avoid spaces and the following characters are not allowed: `[$#*>+/]`
          example: garden
        max_zones:
          type: integer
          description: used to help with validation and avoid errors. This represents the maximum number of Zones that this Garden is able to water
          example: 3
          minimum: 0
        light_schedule:
          type: object
          description: describes when to turn on a light and for how long to leave it on
          properties:
            duration:
              type: string
              format: duration
              description: duration string to determine how long to leave a light on
              example: 14h
            start_time:
              type: string
              format: time
              description: time that the light should be turned on
              example: 23:00:00-07:00
            adhoc_on_time:
              type: string
              format: date-time
              description: if a light delay was used, this persists the time that the light needs to turn back on
          required:
            - duration
            - start_time
      required:
        - max_zones
...